refactor: customHook: useIntersectionCalculation
