+++ src/components/ScatterPlot/SettingsDisplay.jsx
@@ -6,6 +6,8 @@ import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import { v4 as uuidv4 } from 'uuid';
 import { DraggableComponent } from './DraggableComponent';
+import { roundTo } from '../../utils/utils';
+import useIntersectionCalculation from '../../hooks/useIntersectionCalculation';
 
 // eslint-disable-next-line react/prop-types
 function SettingsDisplay({ drawing, setDrawing, data, polygons, setPolygons, colors, setColors }) {
@@ -13,6 +15,12 @@ function SettingsDisplay({ drawing, setDrawing, data, polygons, setPolygons, col
   const colorPickerRefs = useRef({});
   const [selectedPolygons, setSelectedPolygons] = useState([0, 1]);
   const [symbol, setSymbol] = useState('and');
+  const { intersectionCount, intersectionPercentage } = useIntersectionCalculation(
+    polygons,
+    data,
+    selectedPolygons,
+    symbol,
+  );
 
   const handleClickOutside = useCallback(
     (event, polygonId) => {
@@ -42,11 +50,6 @@ function SettingsDisplay({ drawing, setDrawing, data, polygons, setPolygons, col
   const handleDrawing = () => {
     setDrawing(true);
   };
-
-  function roundTo(number, decimalPlaces) {
-    const factor = 10 ** decimalPlaces;
-    return Math.round(number * factor) / factor;
-  }
 
   const handleHide = (polygonId) => {
     setPolygons((prevPolygons) =>
@@ -107,39 +110,6 @@ function SettingsDisplay({ drawing, setDrawing, data, polygons, setPolygons, col
     },
     [setSelectedPolygons],
   );
-
-  const calculateIntersection = (polygon1, polygon2, operator) => {
-    if (!polygon1 || !polygon2) return [];
-    const p1 = polygon1.selected;
-    const p2 = polygon2.selected;
-    if (!p1 || !p2 || p1.length === 0 || p2.length === 0) return [];
-
-    const set1 = new Set(p1.map((point) => JSON.stringify(point)));
-    const set2 = new Set(p2.map((point) => JSON.stringify(point)));
-
-    const result = [];
-
-    if (operator === 'and') {
-      set1.forEach((pointStr) => {
-        if (set2.has(pointStr)) {
-          result.push(JSON.parse(pointStr));
-        }
-      });
-    } else if (operator === 'or') {
-      const combinedSet = new Set([...set1, ...set2]);
-      Array.from(combinedSet).forEach((pointStr) => {
-        result.push(JSON.parse(pointStr));
-      });
-    } else {
-      set1.forEach((pointStr) => {
-        if (!set2.has(pointStr)) {
-          result.push(JSON.parse(pointStr));
-        }
-      });
-    }
-
-    return result;
-  };
 
   const moveEl = useCallback(
     (dragIndex, hoverIndex) => {
@@ -375,26 +345,7 @@ function SettingsDisplay({ drawing, setDrawing, data, polygons, setPolygons, col
               </select>
             </div>
             <p>
-              Calculate:{' '}
-              {
-                calculateIntersection(
-                  polygons.filter((p) => p.id === selectedPolygons[0])[0],
-                  polygons.filter((p) => p.id === selectedPolygons[1])[0],
-                  symbol,
-                ).length
-              }
-              {' / '}
-              {roundTo(
-                (calculateIntersection(
-                  polygons.filter((p) => p.id === selectedPolygons[0])[0],
-                  polygons.filter((p) => p.id === selectedPolygons[1])[0],
-                  symbol,
-                ).length /
-                  data.length) *
-                  100,
-                2,
-              )}
-              %
+              Calculate: {intersectionCount} / {intersectionPercentage}%
             </p>
           </>
         )}
